requirements.txt
# Web & API
Flask==2.3.4
python-dotenv==1.0.0

# Kafka
kafka-python==2.1.0

# PostgreSQL
psycopg2-binary==2.9.9
SQLAlchemy==2.0.22

# Data handling
pandas==2.1.0
numpy==1.26.4

# Dashboard
streamlit==1.30.0
plotly==5.21.0

# Airflow
apache-airflow==2.9.2
apache-airflow-providers-postgres==6.7.1
apache-airflow-providers-apache-kafka==5.2.0

